+++ packages/contract/src/__test__/coverage-contract/coverage-contract.test.ts
@@ -1,15 +1,16 @@
 import { NativeAssetId } from '@fuel-ts/constants';
 import type { BN } from '@fuel-ts/math';
 import { bn, toHex } from '@fuel-ts/math';
-import { Provider, LogReader } from '@fuel-ts/providers';
-import { TestUtils } from '@fuel-ts/wallet';
+import { Provider, ScriptTransactionRequest } from '@fuel-ts/providers';
+import type { Message } from '@fuel-ts/providers';
+import { Wallet, TestUtils } from '@fuel-ts/wallet';
 import { readFileSync } from 'fs';
 import { join } from 'path';
 
 import type Contract from '../../contracts/contract';
 import ContractFactory from '../../contracts/contract-factory';
 
-import abi from './out/debug/coverage-contract-flat-abi.json';
+import abi from './out/debug/coverage-contract-abi.json';
 
 const RUST_U8_MAX = 255;
 const RUST_U16_MAX = 65535;
@@ -234,20 +235,23 @@ describe('Coverage Contract', () => {
   });
 
   it('should test u8 vector input', async () => {
-    const { value, transactionResult } = await contractInstance.functions
+    const { value, logs } = await contractInstance.functions
       .check_u8_vector([1, 2, 3, 4, 5])
       .call();
+
     expect(value).toBeTruthy();
-    const logReader = new LogReader(transactionResult.receipts);
-    expect(logReader.toArray()).toStrictEqual([
+
+    const formattedLog = logs.map((l) => (typeof l === 'string' ? l : l.toNumber()));
+
+    expect(formattedLog).toEqual([
       'vector.buf.ptr',
-      '14464',
+      14464,
       'vector.buf.cap',
-      '5',
+      5,
       'vector.len',
-      '5',
+      5,
       'addr_of vector',
-      '14440',
+      14440,
     ]);
   });
 
@@ -342,5 +346,118 @@ describe('Coverage Contract', () => {
       baz: value.baz,
     };
     expect(unhexed).toStrictEqual(last);
+  });
+
+  it('should get initial state messages from node', async () => {
+    const provider = new Provider('http://127.0.0.1:4000/graphql');
+
+    const WALLET_A = new Wallet(
+      '0x1ff16505df75735a5bcf4cb4cf839903120c181dd9be6781b82cda23543bd242',
+      provider
+    );
+    const WALLET_B = new Wallet(
+      '0x30bb0bc68f5d2ec3b523cee5a65503031b40679d9c72280cd8088c2cfbc34e38',
+      provider
+    );
+
+    const EXPECTED_MESSAGES_A: Message[] = [
+      {
+        amount: bn(1),
+        sender: WALLET_B.address,
+        recipient: WALLET_A.address,
+        data: [8, 7, 6, 5, 4],
+        nonce: bn(1),
+        daHeight: bn(0),
+      },
+    ];
+    const EXPECTED_MESSAGES_B: Message[] = [
+      {
+        amount: bn('12704439083013451934'),
+        sender: WALLET_A.address,
+        recipient: WALLET_B.address,
+        data: [7],
+        nonce: bn('1017517292834129547'),
+        daHeight: bn('3684546456337077810'),
+      },
+    ];
+
+    const aMessages = await WALLET_A.getMessages();
+    const bMessages = await WALLET_B.getMessages();
+
+    expect(aMessages).toStrictEqual(EXPECTED_MESSAGES_A);
+    expect(bMessages).toStrictEqual(EXPECTED_MESSAGES_B);
+  });
+
+  it('should test sending input messages [1]', async () => {
+    const provider = new Provider('http://127.0.0.1:4000/graphql');
+    const request = new ScriptTransactionRequest({ gasLimit: 1000000 });
+
+    const sender = await TestUtils.generateTestWallet(provider, [[1_000, NativeAssetId]]);
+    const receiver = await TestUtils.generateTestWallet(provider);
+
+    const messages: Message[] = [
+      {
+        amount: bn(900),
+        sender: sender.address,
+        recipient: receiver.address,
+        data: [12, 13, 14],
+        nonce: bn(823),
+        daHeight: bn(0),
+      },
+    ];
+
+    request.addMessages(messages);
+    const response = await sender.sendTransaction(request);
+
+    await response.wait();
+    const receiverMessages = await receiver.getMessages();
+
+    expect(receiverMessages).toStrictEqual(messages);
+  });
+
+  it('should test sending input messages [3]', async () => {
+    const provider = new Provider('http://127.0.0.1:4000/graphql');
+    const request = new ScriptTransactionRequest({ gasLimit: 1000000 });
+
+    const sender = await TestUtils.generateTestWallet(provider, [[1_000, NativeAssetId]]);
+    const receiver = await TestUtils.generateTestWallet(provider);
+
+    const messages: Message[] = [
+      {
+        amount: bn(111),
+        sender: sender.address,
+        recipient: receiver.address,
+        data: [11, 11, 11],
+        nonce: bn(100),
+        daHeight: bn(0),
+      },
+      {
+        amount: bn(222),
+        sender: sender.address,
+        recipient: receiver.address,
+        data: [22, 22, 22],
+        nonce: bn(200),
+        daHeight: bn(0),
+      },
+      {
+        amount: bn(333),
+        sender: sender.address,
+        recipient: receiver.address,
+        data: [33, 33, 33],
+        nonce: bn(300),
+        daHeight: bn(0),
+      },
+    ];
+
+    request.addMessages(messages);
+    const response = await sender.sendTransaction(request);
+
+    await response.wait();
+    const receiverMessages = await receiver.getMessages();
+
+    // sort by nonce, messages are not guaranteed in order
+    receiverMessages.sort((a, b) => a.nonce.toNumber() - b.nonce.toNumber());
+
+    expect(receiverMessages).toStrictEqual(messages);
   });
 });