+++ packages/providers/src/index.ts
@@ -3,8 +3,8 @@
 export * from './coin-quantity';
 export * from './coin';
 export * from './provider';
+export * from './message';
 export { default as Provider } from './provider';
 export * from './transaction-request';
 export * from './transaction-response';
 export * from './util';
-export { default as LogReader } from './LogReader';